Extract feature handler into module
