+++ web-executable/ServantUtilities.hs
@@ -1,0 +1,5 @@
+module ServantUtilities where
+  import Control.Monad.Trans.Either
+  import Servant
+
+  type Handler a           = EitherT ServantErr IO a
