+++ web-executable/products/FeaturesAPI.hs
@@ -5,54 +5,33 @@
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
 
-module ProductsAPI
-  ( ProductsAPI
-  , productsAPI
-  , productsServer
+module Products.FeaturesAPI
+  ( FeaturesAPI
+  , FeatureAPI
+  , productsFeatures
+  , productsFeature
   ) where
 
   import Control.Monad.Except (runExceptT)
   import Control.Monad.IO.Class (liftIO)
-  import Control.Monad.Trans.Either
   import Data.Aeson
   import Data.DirectoryTree
   import qualified Data.List        as L
-  import qualified Data.Text        as T
   import Data.Tree (Tree(Node))
   import qualified Features.Feature as F
   import qualified Products.Product as P
   import Servant
   import qualified Servant.Docs     as SD
-  import Models
-
-  data APIProduct = APIProduct { productID :: P.ProductID
-                               , name      :: T.Text
-                               , repoUrl   :: T.Text
-                               } deriving (Show)
+  import ServantUtilities (Handler)
 
   data APIFeature = APIFeature { featureID :: F.FeatureFile
                                , description :: F.Feature
                                }
 
-  type Handler a           = EitherT ServantErr IO a
-
-  type ProductsAPI = "products" :> ( Get '[JSON] [APIProduct]
-                                :<|> Capture "id" P.ProductID :> FeaturesAPI
-                                :<|> Capture "id" P.ProductID :> FeatureAPI
-                                   )
-
   type FeaturesAPI = "features" :> Get '[JSON] DirectoryTree
 
   type FeatureAPI  = "feature"  :> QueryParam "path" F.FeatureFile
                                 :> Get '[JSON] APIFeature
-
-
-  instance ToJSON APIProduct where
-    toJSON (APIProduct prodID prodName prodRepoUrl) =
-      object [ "id" .= prodID
-             , "name"    .= prodName
-             , "repoUrl" .= prodRepoUrl
-             ]
 
   instance ToJSON APIFeature where
     toJSON (APIFeature featID desc) =
@@ -60,22 +39,10 @@ module ProductsAPI
              , "description" .= desc
              ]
 
-  instance SD.ToSample [APIProduct] [APIProduct] where
-    toSample _ = Just $
-      [ APIProduct 1 "monsters" "http://monsters.com/repo.git"
-      , APIProduct 2 "creatures" "ssh://creatures.com/repo.git"
-      ]
-
-  instance SD.ToSample DirectoryTree DirectoryTree where
-    toSample _ = Just featureDirectoryExample
-
   instance SD.ToSample APIFeature APIFeature where
     toSample _ = Just $ APIFeature { featureID = "/features/werewolves/hunting.feature"
                                    , description = featureFileSample
                                    }
-
-  instance SD.ToCapture (Capture "id" P.ProductID) where
-    toCapture _ = SD.DocCapture "id" "Product id"
 
   instance SD.ToParam (QueryParam "path" F.FeatureFile) where
     toParam _ = SD.DocQueryParam
@@ -86,25 +53,8 @@ module ProductsAPI
       "FeatureFile id (relative file path)"
       SD.Normal
 
-  productsServer :: Server ProductsAPI
-  productsServer = products
-              :<|> productsFeatures
-              :<|> productsFeature
-
-  productsAPI :: Proxy ProductsAPI
-  productsAPI = Proxy
-
-  products :: Handler [APIProduct]
-  products = do
-    prods <- liftIO P.findProducts
-    return $ map toProduct prods
-      where
-        toProduct dbProduct = do
-          let dbProd   = P.toProduct dbProduct
-          let dbProdID = P.toProductID dbProduct
-          APIProduct { productID = dbProdID
-                     , name      = productName dbProd
-                     , repoUrl   = productRepoUrl dbProd }
+  instance SD.ToSample DirectoryTree DirectoryTree where
+    toSample _ = Just featureDirectoryExample
 
   productsFeatures :: P.ProductID -> Handler DirectoryTree
   productsFeatures prodID = do
